refactor(core): upper titles and buttons for TT/TR
New models might not have titles or button labels in
all uppercase as TT and TR do so the translation strings
must be written in normal case.
Instead we uppercases the strings during generation.
In the future, we might want to get rid of separated
translated_string.rs and instead modify the font data.

[no changelog]

refactor(core): upper titles and buttons for TT/TR

WIP

New models might not have titles or button labels in
all uppercase as TT and TR do so the translation strings
must be written in normal case.
Instead we uppercases the strings during generation.
In the future, we might want to get rid of separated
translated_string.rs and instead modify the font data.

[no changelog]