fix(tests): python >=3.11 compatibility for identifiers
Python 3.11 changed IntEnum.__str__ to return the number instead of the
enum value name. This breaks fixtures.json because pytest uses
str(value) to generate the test identifier names, and in a lot of places
our identifiers use the enum values.

This override of `_idval_from_value` explicitly generates a name from
the IntEnum instead of using the __str__ implementation.